move axios configurations to another file
